pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-nanopages"
description = "Generate Django pages from Markdown, HTML, and Django template files."
dynamic = ["version"]
authors = [
    { name="Richard Terry", email="code@radiac.net" },
]
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
keywords = ["django"]
requires-python = ">=3.8"
dependencies = [
    "markdown",
]

[project.optional-dependencies]
full = [
    "pyyaml",
]

[project.entry-points.nanodjango]
django_nanopages = "django_nanopages.nanodjango"

[project.urls]
Homepage = "https://radiac.net/projects/django-nanopages/"
Documentation = "https://django-nanopages.readthedocs.io/en/latest/"
Changelog = "https://django-nanopages.readthedocs.io/en/latest/changelog.html"
Repository = "https://github.com/radiac/django-nanopages"
Issues = "https://github.com/radiac/django-nanopages/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["django_nanopages*"]
exclude = ["docs*", "tests*", "examples*", "src*", "dist*"]

[tool.setuptools.dynamic]
version = {attr = "django_nanopages.__version__"}

[tool.pytest.ini_options]
addopts = "--cov=django_nanopages --cov-report=term --cov-report=html"
testpaths = [
    "tests",
    "django_nanopages",
]
pythonpath = ["."]
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.coverage.run]
source = ["django_nanopages"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = "\\.pyi?$"

[tool.isort]
multi_line_output = 3
line_length = 88
known_django = "django"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
lines_after_imports = 2
skip = [".git"]

[tool.mypy]
follow_imports = "skip"
ignore_missing_imports = true

[tool.doc8]
max-line-length = 88
ignore-path = ["*.txt"]

[tool.ruff]
line-length = 88
lint.select = ["E", "F"]
lint.ignore = [
    "E501",  # line length
]
exclude = [
    ".git",
    "dist",
]
